+++ app.js
@@ -1,9 +1,10 @@
 const path = require('path');
 const express = require('express');
 const bodyParser = require('body-parser');
+const expressLayouts = require('express-ejs-layouts');
 
 //importando os arquivos de rotas
-const adminData = require('./routes/admin');
+const adminRoutes = require('./routes/admin');
 const routesShop = require('./routes/shop');
 
 const app = express();
@@ -11,6 +12,7 @@ const app = express();
 //configurando o template engine
 app.set('view engine', 'ejs');
 app.set('views', 'views');
+app.use(expressLayouts);
 
 //configurando o bodyparser
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -19,7 +21,7 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, 'public')));
 
 //importando as rotas
-app.use('/admin', adminData.routes);
+app.use('/admin', adminRoutes);
 app.use(routesShop);
 
 //pagina 404
