+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "body-parser": "^1.19.0",
     "ejs": "^3.1.5",
-    "express": "^4.16.3"
+    "express": "^4.16.3",
+    "express-ejs-layouts": "^2.5.0"
   }
 }
