+++ routes/admin.js
@@ -1,26 +1,9 @@
-const path = require('path');
 const express = require('express');
-const rootDir = require('../util/path');
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-const products = [];
+router.get('/add-product', productsController.getAddProduct);
+router.post('/add-product', productsController.postAddProduct);
 
-router.get('/add-product', (req, res, next) => {
-    //res.sendFile(path.join(rootDir, 'views', 'add-product.html'));
-    res.render('add-product', {
-        pageTitle: 'Add Product',
-        path: '/admin/add-product',
-        formsCSS: true,
-        productCSS: true,
-        activeAddProduct: true
-    });
-});
-
-router.post('/add-product', (req, res, next) => {
-    products.push({ title: req.body.title });
-    res.redirect('/');
-});
-
-exports.routes = router;
-exports.products = products;
+module.exports = router;
