+++ routes/shop.js
@@ -1,22 +1,8 @@
-const path = require('path');
 const express = require('express');
-const rootDir = require('../util/path');
-const adminData = require('./admin');
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-    //console.log('shop page', adminData.products);
-    //res.sendFile(path.join(rootDir, 'views', 'shop.html'));
-    const products = adminData.products;
-    res.render('shop', {
-      prods: products,
-      pageTitle: 'Shop',
-      path: '/',
-      hasProducts: products.length > 0,
-      activeShop: true,
-      productCSS: true
-    });
-});
+router.get('/', productsController.getProduct);
 
 module.exports = router;
