+++ views/add-product.ejs
@@ -1,19 +1,9 @@
-<%- include('includes/head.ejs') %>
-    <link rel="stylesheet" href="/css/forms.css">
-    <link rel="stylesheet" href="/css/product.css">
-</head>
+<%- contentFor('body') %>
+<form class="product-form" action="/admin/add-product" method="POST">
+    <div class="form-control">
+        <label for="title">Title</label>
+        <input type="text" name="title" id="title">
+    </div>
 
-<body>
-   <%- include('includes/navigation.ejs') %>
-
-    <main>
-        <form class="product-form" action="/admin/add-product" method="POST">
-            <div class="form-control">
-                <label for="title">Title</label>
-                <input type="text" name="title" id="title">
-            </div>
-
-            <button class="btn" type="submit">Add Product</button>
-        </form>
-    </main>
-<%- include('includes/end.ejs') %>
+    <button class="btn" type="submit">Add Product</button>
+</form>
