+++ views/html/shop.html
