+++ views/html/teste.html
@@ -1,0 +1,3 @@
+<turbo-frame id="new_message">
+    <h1>Pedaçõ!</h1>
+</turbo-frame>
